requirements.txt
click==8.1.3
deepface==0.0.79
dlib==19.24.0
face_recognition_models==0.3.0
Flask==2.2.2
Flask_DebugToolbar==0.13.1
Flask_SocketIO==5.3.5
numpy==1.24.1
opencv_contrib_python==4.7.0.68
opencv_python==4.7.0.68
Pillow==10.0.0
scikit_learn==1.2.2
setuptools==58.1.0
waitress==2.1.2
